Moved classes to relion.objects and compute data in protocol
